+++ lib/modules/state/state_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: state
-// Updated at: 2026-05-15 21:09:29
-// Commit: feat(state): integrate custom telemetry logger in state
+// Updated at: 2026-05-16 12:53:57
+// Commit: fix(state): prevent double tap duplicate action in state
 
 class stateComponentState {
-  final String commitRef = '1197';
-  final String lastMessage = 'feat(state): integrate custom telemetry logger in state';
-  final DateTime timestamp = DateTime.parse('2026-05-15T21:09:29+05:00');
+  final String commitRef = '1205';
+  final String lastMessage = 'fix(state): prevent double tap duplicate action in state';
+  final DateTime timestamp = DateTime.parse('2026-05-16T12:53:57+05:00');
 }